requirements.txt
#  Django Core 
Django 5.1.2
djangorestframework 3.15.2
django-cors-headers 4.5.0
python-decouple 3.8
django-filter 24.2

#  PostgreSQL
psycopg2-binary 2.9.10

#  Autenticación y Seguridad 
djangorestframework-simplejwt 5.4.0
#  Documentación (Swagger / API Docs) 
drf-yasg 1.21.7
coreapi 2.3.3
coreschema 0.0.4

Pillow 11.0.0

requests 2.32.3
python-dotenv 1.0.1

#  IA y Reportes (para ciclos futuros) 
pandas 2.2.3
numpy 2.1.3
scikit-learn 1.5.2
joblib 1.4.2
reportlab 4.2.5
openpyxl 3.1.5
weasyprint 63.1

stripe 10.6.0

pydub 0.25.1
SpeechRecognition 3.10.1

setuptools 70.1.1
wheel 0.44.0

#para reportes por voz 
SpeechRecognition 3.10.4
ffmpeg-python 0.2.0
